refactor(search): delegate condition building to SqlQueryBuilder
